src/requirements.txt
uvicorn[standard]==0.30.0
uvicorn-worker==0.2.0
gunicorn==22.0.0
Django==5.0.6
psycopg==3.1.19
psycopg_binary==3.1.19
whitenoise==6.6.0
django_components==0.81
django-debug-toolbar==4.4.2
django-htmx==1.17.3
django-allauth[socialaccount]==0.63.2
django-widget-tweaks==1.5.0